# Assert Zeros option for AIRs