--- a/air/src/air.rs
+++ b/air/src/air.rs
@@ -110,9 +110,7 @@ pub trait AirBuilder: Sized {
 
     /// Assert that a given array consists of only boolean values.
     fn assert_bools<const N: usize, I: Into<Self::Expr>>(&mut self, array: [I; N]) {
-        let zero_array = array.map(|x| {
-            x.into().bool_check()
-        });
+        let zero_array = array.map(|x| x.into().bool_check());
         self.assert_zeros(zero_array);
     }