--- a/air/src/utils.rs
+++ b/air/src/utils.rs
@@ -119,7 +119,7 @@ pub fn add3<AB: AirBuilder>(
     let acc_32 = a[1] - b[1] - c[1].clone() - d[1].clone();
     let acc = acc_16.clone() + AB::Expr::from_prime_subfield(two_16) * acc_32;
 
-    builder.assert_zeroes([
+    builder.assert_zeros([
         acc.clone()
             * (acc.clone() + AB::Expr::from_prime_subfield(two_32))
             * (acc + AB::Expr::from_prime_subfield(two_32.double())),
@@ -179,7 +179,7 @@ pub fn add2<AB: AirBuilder>(
     let acc_32 = a[1] - b[1] - c[1].clone();
     let acc = acc_16.clone() + AB::Expr::from_prime_subfield(two_16) * acc_32;
 
-    builder.assert_zeroes([
+    builder.assert_zeros([
         acc.clone() * (acc + AB::Expr::from_prime_subfield(two_32)),
         acc_16.clone() * (acc_16 + AB::Expr::from_prime_subfield(two_16)),
     ]);
@@ -216,5 +216,5 @@ pub fn xor_32_shift<AB: AirBuilder>(
 
     // As both b and c have been range checked to be boolean, all the (b ^ (c << shift))
     // are also boolean and so this final check additionally has the effect of range checking a[0], a[1].
-    builder.assert_zeroes([a[0] - sum_0_16, a[1] - sum_16_32]);
+    builder.assert_zeros([a[0] - sum_0_16, a[1] - sum_16_32]);
 }