--- a/baby-bear/src/aarch64_neon/packing.rs
+++ b/baby-bear/src/aarch64_neon/packing.rs
@@ -29,7 +29,8 @@ mod tests {
 
     test_packed_field!(
         crate::PackedBabyBearNeon,
-        crate::PackedBabyBearNeon::ZERO,
+        &[crate::PackedBabyBearNeon::ZERO],
+        &[crate::PackedBabyBearNeon::ONE],
         p3_monty_31::PackedMontyField31Neon::<crate::BabyBearParameters>(super::SPECIAL_VALS)
     );
 }