--- a/baby-bear/src/x86_64_avx2/packing.rs
+++ b/baby-bear/src/x86_64_avx2/packing.rs
@@ -11,6 +11,8 @@ const WIDTH: usize = 8;
 
 impl MontyParametersAVX2 for BabyBearParameters {
     const PACKED_P: __m256i = unsafe { transmute::<[u32; WIDTH], _>([0x78000001; WIDTH]) };
+    const PACKED_P_HIGH: __m256i =
+        unsafe { transmute::<[[u32; 2]; WIDTH / 2], _>([[0x78000001, 0]; WIDTH / 2]) };
     const PACKED_MU: __m256i = unsafe { transmute::<[u32; WIDTH], _>([0x88000001; WIDTH]) };
 }