--- a/baby-bear/src/x86_64_avx2/packing.rs
+++ b/baby-bear/src/x86_64_avx2/packing.rs
@@ -28,7 +28,8 @@ mod tests {
 
     test_packed_field!(
         crate::PackedBabyBearAVX2,
-        crate::PackedBabyBearAVX2::ZERO,
+        &[crate::PackedBabyBearAVX2::ZERO],
+        &[crate::PackedBabyBearAVX2::ONE],
         p3_monty_31::PackedMontyField31AVX2::<crate::BabyBearParameters>(super::SPECIAL_VALS)
     );
 }