--- a/bn254-fr/benches/bench_field.rs
+++ b/bn254-fr/benches/bench_field.rs
@@ -1,19 +1,35 @@
 use criterion::{Criterion, criterion_group, criterion_main};
 use p3_bn254_fr::Bn254Fr;
-use p3_field_testing::bench_func::{
-    benchmark_add_latency, benchmark_add_throughput, benchmark_inv, benchmark_iter_sum,
-    benchmark_sub_latency, benchmark_sub_throughput,
+use p3_field_testing::{
+    bench_func::{
+        benchmark_add_latency, benchmark_add_throughput, benchmark_inv, benchmark_iter_sum,
+        benchmark_sub_latency, benchmark_sub_throughput,
+    },
+    benchmark_tree_sum,
 };
 
 type F = Bn254Fr;
 
 fn bench_field(c: &mut Criterion) {
     let name = "BN254Fr";
-    const REPS: usize = 1000;
+    const REPS: usize = 100;
     benchmark_inv::<F>(c, name);
+    benchmark_iter_sum::<F, 1, REPS>(c, name);
+    benchmark_tree_sum::<F, 1, REPS>(c, name);
+    benchmark_iter_sum::<F, 2, REPS>(c, name);
+    benchmark_tree_sum::<F, 2, REPS>(c, name);
+    benchmark_iter_sum::<F, 3, REPS>(c, name);
+    benchmark_tree_sum::<F, 3, REPS>(c, name);
     benchmark_iter_sum::<F, 4, REPS>(c, name);
-    benchmark_iter_sum::<F, 8, REPS>(c, name);
-    benchmark_iter_sum::<F, 12, REPS>(c, name);
+    benchmark_tree_sum::<F, 4, REPS>(c, name);
+    // benchmark_iter_sum::<F, 8, REPS>(c, name);
+    // benchmark_tree_sum::<F, 8, REPS>(c, name);
+    // benchmark_iter_sum::<F, 12, REPS>(c, name);
+    // benchmark_tree_sum::<F, 12, REPS>(c, name);
+    // benchmark_iter_sum::<F, 16, REPS>(c, name);
+    // benchmark_tree_sum::<F, 16, REPS>(c, name);
+    // benchmark_iter_sum::<F, 64, REPS>(c, name);
+    // benchmark_tree_sum::<F, 64, REPS>(c, name);
 
     // Note that each round of throughput has 10 operations
     // So we should have 10 * more repetitions for latency tests.