--- a/bn254-fr/benches/bench_field.rs
+++ b/bn254-fr/benches/bench_field.rs
@@ -1,12 +1,10 @@
 use criterion::{Criterion, criterion_group, criterion_main};
 use p3_bn254_fr::Bn254Fr;
-use p3_field_testing::{
-    bench_func::{
-        benchmark_add_latency, benchmark_add_throughput, benchmark_inv, benchmark_iter_sum,
-        benchmark_sub_latency, benchmark_sub_throughput,
-    },
-    benchmark_tree_sum,
+use p3_field_testing::bench_func::{
+    benchmark_add_latency, benchmark_add_throughput, benchmark_inv, benchmark_iter_sum,
+    benchmark_sub_latency, benchmark_sub_throughput,
 };
+use p3_field_testing::benchmark_tree_sum;
 
 type F = Bn254Fr;