--- a/bn254-fr/benches/bench_field.rs
+++ b/bn254-fr/benches/bench_field.rs
@@ -4,7 +4,7 @@ use p3_field_testing::bench_func::{
     benchmark_add_latency, benchmark_add_throughput, benchmark_inv, benchmark_iter_sum,
     benchmark_sub_latency, benchmark_sub_throughput,
 };
-use p3_field_testing::benchmark_tree_sum;
+use p3_field_testing::benchmark_sum_array;
 
 type F = Bn254Fr;
 
@@ -14,9 +14,9 @@ fn bench_field(c: &mut Criterion) {
     benchmark_inv::<F>(c, name);
 
     benchmark_iter_sum::<F, 4, REPS>(c, name);
-    benchmark_tree_sum::<F, 4, REPS>(c, name);
+    benchmark_sum_array::<F, 4, REPS>(c, name);
     benchmark_iter_sum::<F, 64, REPS>(c, name);
-    benchmark_tree_sum::<F, 64, REPS>(c, name);
+    benchmark_sum_array::<F, 64, REPS>(c, name);
 
     // Note that each round of throughput has 10 operations
     // So we should have 10 * more repetitions for latency tests.