This pull request is composed of three main parts.

- Introduce a new method to the trait `AirBuilder` called `assert_zeros`. This allows constraints which of the form `builder.assert_zeros(arr)` for an array `arr`. We also add `assert_bools` for convenience. Currently I've only updated the `KECCAK-air` to use this new method. I'll update the `Blake-3-air` and `Poseidon2-airs` in different PR's. The trait has a generic implementation of this but a specialized one is given in `uni-stark/src/folder.rs`. This new method can take advantage of fast dot product code.
- Introduce a new methods to `PrimeCharacteristicRing` and `PackedField` called `tree_sum` and `dot_product_scalar_packed` respectively. The first method sums a given slice of elements using a tree like structure to minimize dependency chains. We need this as the compiler does not realise that addition is associative. The second method computes a dot product between a slice of `PackedField` elements and `Field` elements. I've written AVX2 and AVX512 implementations for this second method over `MontyField31`'s. These make use of delayed reduction techniques are significantly faster (1.5x - 3x) than the naive approach to dot products. (I also remove trit_check as we never actually use this).
- Refactor the testing crate. Currently Field and PackedFields do very similar tests but with different code. I refactored the testing crate to begin to mitigate this and use the same code for both fields. This lets us test the implementations of `tree_sum` and `dot_product_scalar_packed` (as well as other code in `PrimeCharacteristicRing`) more easily and helps ensure the test suite is comprehensive in both cases.

The main point of this PR is that this should significantly speed up constraint evaluation (roughly measured by `compute quotient polynomial` in our trace).

We benchmark on AVX2/AVX512 using the commands:

`cargo run --example prove_prime_field_31 --release --features parallel -- --field koala-bear --objective keccak-f-permutations --log-trace-length 17 --discrete-fourier-transform radix-2-dit-parallel --merkle-hash keccak-f`

`cargo +nightly run --features nightly-features --example prove_prime_field_31 --release --features parallel -- --field koala-bear --objective keccak-f-permutations --log-trace-length 17 --discrete-fourier-transform radix-2-dit-parallel --merkle-hash keccak-f`

On my cpu with znver4: (Take the AVX2 numbers with a grain of salt, something strange is happening in compilation on znver4)

- Branch: AVX2:   compute quotient polynomial [ 156ms | 7.69% ]
- Branch: AVX512: compute quotient polynomial [ 200ms | 9.62% ]

- Main: AVX2:   compute quotient polynomial [ 183ms | 8.91% ]
- Main: AVX512: compute quotient polynomial [ 260ms | 11.69% ]

On my cpu with alderlake (only has AVX-2): 

- Branch: AVX2:   compute quotient polynomial [ 234ms | 7.98% ]
- Main: AVX2:   compute quotient polynomial [ 286ms | 9.76% ]