@SyxtonPrime 2025-03-10
> Um, could this not have been three separate PRs? If not, and the three things you outline in the description are fundamentally intertwined, could you add a few words about how?

I did think about separating this into 3 PR's the reason I ended up combining it was:
- Adding `assert_zeros` to `AirBuilder` without any of the dot product improvements would have no effect (It would just loop through the given array and call `assert_zero` on each). Thus is doesn't really make sense to merge `assert_zeros` without merging `tree_sum` and `dot_product_scalar_packed`.
- Similarly, without `assert_zeros`, `tree_sum` and `dot_product_scalar_packed` would be unused. They do have a couple of other potential applications in the code base but those will involve some restructuring. Hence merging `tree_sum` and `dot_product_scalar_packed` without `assert_zeros` would broadly add to code bloat without a clear reason why.
- Finally, if we are going to merge `tree_sum` and `dot_product_scalar_packed`, we definitely want to add tests for them. These tests ended up being a little painful to write and once I got them working it was easy to just migrate over some of the other tests.

I could probably split off some of the testing changes but it wouldn't really reduce the number of lines of code added by that much as adding `test_dot_product` and `test_sums` make up a large part of those changes.